Slight refactor, add Associated Devices for measurement sets. Add Measurement Component
