+++ src/components/measurement-item/measurement-item.scss
@@ -1,0 +1,37 @@
+.item-card {
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  margin: 5px 10px 0px 10px;
+  border-radius: 5px;
+  background-color: var(--ion-color-tertiary-tint);
+}
+
+.item-image {
+  margin: 5px;
+  border: dashed;
+  height: 60px;
+  width: 60px;
+}
+
+.card-details {
+  width: 55%;
+  margin: 10px;
+}
+
+.detail-text {
+  font-weight: 300;
+  margin: 0px 2px 0px 2px;
+}
+
+.detail-title {
+  margin: 5px;
+  font-weight: 500;
+}
+
+.item-date {
+  margin-top: 15px;
+  margin-right: 5px;
+  font-weight: 300;
+  font-size: 14px;
+}
