+++ src/core/stores/measurements.ts
@@ -19,6 +19,7 @@ const makeMockMeasurements = (n: number) =>
       createdAt: faker.date.recent(),
       modifiedAt: faker.date.recent(),
       data: [],
+      associatedDevices: [],
     };
   }, n);
 //</editor-fold>
