+++ src/core/stores/profiles.ts
@@ -34,6 +34,7 @@ let mockArm: MeasurementSet = {
       units: "cm",
     },
   ],
+  associatedDevices: [{ device: mockDevices()[0], notes: "My Device" }],
 };
 const mockProfiles = (): Profile[] => [
   {
